MFC r351819: procstat/tests: Fix flakiness by waiting for program to start
Some of the procstat tests start a program "while1" and examine the process
using procstat, but did not wait properly for it to start (kill -0 will
succeed immediately after the child process has been created).

Instead, have "while1" write something when it starts, and use a fifo to
wait for that.

PR:		233587, 233588


git-svn-id: https://svn.freebsd.org/base/stable/11@352492 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f