Merge from stable/12 r352735 and r352741
Merge all the stability fixes for the mpr and mps drivers. This fixes a
number of different panics. Unfortunately, mps now requires atomic_swap_64
to work properly, so it has been disabled on 32-bit powerpc and mips. The
impact should be negligible, however, since this device is difficult to
attach to those platforms.

Approved by:	re@ (glen)
Relnotes:	YES (for ppc32 removal)


git-svn-id: https://svn.freebsd.org/base/releng/12.1@352761 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f