MFC r352804:
Correct the final argument name in the top(1) manpage.

The description talks about 'number', while the final argument was
'count'.  Since 'count' is already used for the count of displays,
change the final argument name to 'number'.

MFC r352818:

Make fractional delays for top(1) work for interactive mode.

In r334906, the -s option was changed to allow fractional times, but
this only functioned correctly for batch mode.  In interactive mode, any
delay below 1.0 would get floored to zero.  This would put top(1) into a
tight loop, which could be difficult to interrupt.

Fix this by storing the -s option value (after validation) into a struct
timeval, and using that struct consistently for delaying with select(2).

Next up is to allow interactive entry of a fractional delay value.

MFC r352819:

Allow entering fractional delays in top(1) interactive mode.

This uses the same logic as with the -s option, first validating the
entered value, then storing the result in a struct timeval.


git-svn-id: https://svn.freebsd.org/base/stable/12@352889 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f