MFC r352863: fdt_slicer: bump to SI_ORDER_THIRD following r347183
r347183 bumped GEOM classes to SI_ORDER_SECOND to resolve a race between
them and the initialization of devsoftc.mtx in devinit, but missed this
dependency on g_flashmap that may now lose the race against GEOM
classes/g_init.

There's a great comment that describes the situation that has also been
updated with the new ordering of GEOM classes.


git-svn-id: https://svn.freebsd.org/base/stable/12@353041 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f