MFC r352220:
Avoid the use of the non-portable -D argument to ls.

This was used to store the mtime of the source file in a commment in a
generated header file.  This is of little-to-no diagnostic value and
the result doesn't even end up in the source tree.

Reported by:	arichardson
Reviewed by:	arichardson
Sponsored by:	DARPA, AFRL
Differential Revision:	https://reviews.freebsd.org/D21605


git-svn-id: https://svn.freebsd.org/base/stable/12@353055 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f