MFC r319795:
Disable the -O (output fields) option in poollist() (ippool -l) for
now. The option does not presently work. However, similar functions in
ipfstat (for state) and ipnat (for nat) do work and provide outputs that
can be easily parsed by shell scripts or subsequently loaded into CSV
files.  The intention here is to return to this option to make it work.
I suspect the problem is in printpoolfields.c.


git-svn-id: https://svn.freebsd.org/base/stable/11@353076 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f