MFC r353145:
Plumb an mbuf leak in a code path that should not be taken. Also avoid
that this path is taken by setting the tail pointer correctly.
There is still bug related to handling unordered unfragmented messages
which were delayed in deferred handling.
This issue was found by OSS-Fuzz testing the usrsctp stack and reported
in
https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=17794


git-svn-id: https://svn.freebsd.org/base/stable/12@353401 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f