MFC r353443
mountroot: run statfs after mounting devfs

The usual flow for mounting a file system is to VFS_MOUNT() and then
immediately VFS_STATFS().

That's not done in vfs_mountroot_devfs(), which means the
mp->mnt_stat.f_iosize field is not correctly populated, which in turn
causes us to mark valid aio operations as unsafe (because the io size is
set to 0), ultimately causing the aio_test:md_waitcomplete test to fail.

Sponsored by:	Axiado


git-svn-id: https://svn.freebsd.org/base/stable/12@353716 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f